backend/requirements.txt
# Core Django
Django>=5.1.3
django-cors-headers>=4.6.0
django-filter>=24.3
django-extensions>=3.2.3
whitenoise==6.9.0
psycopg2-binary>=2.9.9

# API
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1

# Tasks & Background Processing
celery>=5.4.0
redis>=5.0.0
django-redis>=5.4.0

# Production Server
gunicorn==23.0.0

# Integration & AI
google-genai>=0.1.0
python-dotenv==1.0.1
requests>=2.32.5

# Utilities
pillow>=11.0.0
